refactor: storage client implementations to use unified Tauri commands
- Removed direct invoke calls in storage clients and replaced them with a unified command structure.
- Introduced `invokeListDirectory`, `invokeDownloadFile`, and other command wrappers in `BaseStorageClient`.
- Updated `HuggingFaceStorageClient`, `LocalStorageClient`, `OSSStorageClient`, `WebDAVStorageClient`, and `StorageManager` to utilize the new command structure.
- Simplified error handling by checking command results for errors.
- Enhanced type safety and consistency across storage operations.
- Removed unused timeout configurations and related logic.
- Updated type definitions to align with backend changes, including adjustments to `FileContent` and `FilePreview`.
- Added utility functions for safe type conversions and file size calculations.

+++ src-tauri/src/archive/formats/mod.rs
@@ -21,7 +21,7 @@ pub trait CompressionHandlerDispatcher: Send + Sync {
         client: Arc<dyn StorageClient>,
         file_path: &str,
         filename: &str,
-        max_size: Option<usize>,
+        max_size: Option<u32>,
     ) -> Result<ArchiveInfo, String>;
 
     /// 通过存储客户端提取文件预览（统一接口，支持流式提取、进度回调和取消信号）
